+++ nengo/networks/__init__.py
@@ -1,4 +1,5 @@
 from .actionselection import BasalGanglia, Thalamus
+from .assoc_mem import AssociativeMemory
 from .circularconvolution import CircularConvolution
 from .ensemblearray import EnsembleArray
 from .integrator import Integrator